requirements.txt
python-docx
python-multipart
fastapi
jinja2
confuse
pandas
openpyxl